requirements.txt
fastapi
uvicorn[standard]
streamlit
pymongo[srv]
python-dotenv
google-generativeai
faiss-cpu
numpy
pydantic
httpx
